+++ AdminWebsite/AdminWebsite/ClientApp/package.json
@@ -6,7 +6,7 @@
     "start": "ng serve",
     "build": "ng build",
     "build-prod": "ng build --prod",
-    "test": "ng test --code-coverage",
+    "test": "ng test --code-coverage --source-map",
     "test-once": "ng test --watch=false --code-coverage",
     "test-once-ci": "ng test --browsers=ChromeHeadlessNoPrompt --watch=false --code-coverage",
     "lint": "ng lint --fix && npm run prettify-fix",
