+++ AdminWebsite/AdminWebsite/ClientApp/src/app/app.module.ts
@@ -47,6 +47,8 @@ import { VhoWorkHoursTableComponent } from './work-allocation/edit-work-hours/vh
 import { VhoWorkHoursNonAvailabilityTableComponent } from './work-allocation/edit-work-hours/vho-work-hours-non-availability-table/vho-work-hours-non-availability-table.component';
 import { ConfirmDeleteHoursPopupComponent } from './popups/confirm-delete-popup/confirm-delete-popup.component';
 import { UnallocatedHearingsComponent } from './dashboard/unallocated-hearings/unallocated-hearings.component';
+import { UploadWorkHoursComponent } from './work-allocation/upload-work-hours/upload-work-hours.component';
+import { ManageTeamComponent } from './work-allocation/manage-team/manage-team.component';
 
 export function loadConfig(configService: ConfigService) {
     return () => configService.loadConfig();
@@ -68,7 +70,9 @@ export function loadConfig(configService: ConfigService) {
         VhoWorkHoursTableComponent,
         VhoWorkHoursNonAvailabilityTableComponent,
         ConfirmDeleteHoursPopupComponent,
-        UnallocatedHearingsComponent
+        UnallocatedHearingsComponent,
+        UploadWorkHoursComponent,
+        ManageTeamComponent
     ],
     imports: [
         MomentModule,
