+++ AdminWebsite/AdminWebsite/ClientApp/src/app/work-allocation/manage-team/manage-team.component.html
@@ -1,0 +1,16 @@
+<details class="govuk-details" data-module="govuk-details">
+  <summary class="govuk-details__summary govuk-!-margin-bottom-2" tabindex="0" role="button">
+    <span class="govuk-details__summary-text" id="manage-team"> Manage team </span>
+  </summary>
+  <div class="govuk-form-group govuk-!-margin-left-6">
+    <div class="govuk-form-group">
+      <label class="govuk-label govuk-!-font-weight-bold govuk-!-margin-bottom-2" for="search-team-member"> Search for team member </label>
+      <div class="govuk-grid-row">
+        <div class="govuk-grid-column-one-third">
+          <input class="govuk-input" id="search-team-member" name="search-team-member" type="text" />
+        </div>
+        <button class="govuk-button" data-module="govuk-button">Search</button>
+      </div>
+    </div>
+  </div>
+</details>
