+++ AdminWebsite/AdminWebsite/ClientApp/src/app/work-allocation/manage-team/manage-team.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-manage-team',
+    templateUrl: './manage-team.component.html',
+    styleUrls: ['./manage-team.component.scss']
+})
+export class ManageTeamComponent implements OnInit {
+    constructor() {}
+
+    ngOnInit(): void {}
+}
